requirements.txt
pip>=25.2
setuptools>=70
wheel>=0.43Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
email-validator==2.1.0
Werkzeug==3.0.1
Pillow==10.1.0
Flask-Mail==0.9.1
pytz==2024.1
pandas==2.1.4
openpyxl==3.1.2
gunicorn==21.2.0
Flask-Migrate==4.0.5